# swagger.yaml
swagger: '2.0'
info:
  title: Notes Service API
  version: '1.0.0'
  description: A simple API for managing notes
host: localhost:3000
basePath: /
schemes:
  - http
paths:
  /notes/{noteName}:
    get:
      summary: Retrieve a note by name
      parameters:
        - name: noteName
          in: path
          required: true
          type: string
      responses:
        200:
          description: Returns the note text
        404:
          description: Note not found
    put:
      summary: Update an existing note
      parameters:
        - name: noteName
          in: path
          required: true
          type: string
        - name: body
          in: body
          required: true
          schema:
            type: string
      responses:
        200:
          description: Note updated successfully
        404:
          description: Note not found
    delete:
      summary: Delete a note
      parameters:
        - name: noteName
          in: path
          required: true
          type: string
      responses:
        200:
          description: Note deleted successfully
        404:
          description: Note not found
  /notes:
    get:
      summary: List all notes
      responses:
        200:
          description: Returns a list of notes
          schema:
            type: array
            items:
              type: object
              properties:
                name:
                  type: string
                text:
                  type: string
  /write:
    post:
      summary: Create a new note
      consumes:
        - multipart/form-data
      parameters:
        - name: note_name
          in: formData
          required: true
          type: string
        - name: note
          in: formData
          required: true
          type: string
      responses:
        201:
          description: Note created successfully
        400:
          description: Note already exists
  /UploadForm.html:
    get:
      summary: Serve the upload form
      responses:
        200:
          description: Returns the HTML upload form